refactor LLM to dedicated folder
