+++ automata/cli/scripts/run_code_embedding.py
@@ -7,7 +7,7 @@ from automata.config.config_types import ConfigCategory
 from automata.core.base.database.vector import JSONVectorDatabase
 from automata.core.coding.py.module_loader import py_module_loader
 from automata.core.embedding.code_embedding import SymbolCodeEmbeddingHandler
-from automata.core.embedding.embedding_types import OpenAIEmbedding
+from automata.core.llm.providers.openai import OpenAIEmbedding
 from automata.core.symbol.graph import SymbolGraph
 from automata.core.symbol.symbol_utils import get_rankable_symbols
 from automata.core.utils import config_fpath
