+++ automata/core/embedding/doc_embedding.py
@@ -7,10 +7,9 @@ from jinja2 import Template
 from automata.config.prompt.docs import DEFAULT_DOC_GENERATION_PROMPT
 from automata.core.base.database.vector import VectorDatabaseProvider
 from automata.core.context.py.retriever import PyContextRetriever
+from automata.core.llm.embedding import EmbeddingProvider, SymbolEmbeddingHandler
 from automata.core.symbol.search.symbol_search import SymbolSearch
 from automata.core.symbol.symbol_types import Symbol, SymbolDocEmbedding
-
-from .embedding_types import EmbeddingProvider, SymbolEmbeddingHandler
 
 logger = logging.getLogger(__name__)
