+++ automata/core/llm/completion.py
